.autonomous_system/scripts/trigger_self_construction.py
"""
TRIGGER SELF CONSTRUCTION
This script uses the 'CodeGenerator' module to autonomously build the 'TrueAIAgent' class.
It demonstrates the system's ability to help build itself.
"""

import asyncio
from pathlib import Path
from AUTONOMOUS_SYSTEM.code_generator import CodeGenerator

async def run_self_construction():
    print("🤖 SYSTEM AWAKENING: Initiating Self-Construction Protocol...")
    
    # 1. Initialize the Elite Code Generator
    project_root = Path(__file__).parent
    generator = CodeGenerator(project_root)
    
    print("   ...Code Generator Online")
    
    # 2. Define the Requirement for the AI Agent
    class_name = "TrueAIAgent"
    description = "A professional-grade AI Agent capable of autonomous decision making, task execution, and learning."
    
    methods = [
        {
            "name": "process_request",
            "type": "async",
            "doc": "Process a complex user request asynchronously and return a structured response."
        },
        {
            "name": "learn_from_interaction",
            "type": "async",
            "doc": "Analyze the result of an interaction and update internal knowledge base."
        },
        {
            "name": "execute_tool",
            "type": "async",
            "doc": "Execute a specific tool or command with safety checks."
        },
        {
            "name": "get_status",
            "type": "sync",
            "doc": "Return the current health and status of the agent."
        }
    ]
    
    # 3. Command the System to Generate Code
    print(f"   ...Generating source code for {class_name}...")
    source_code = generator.generate_optimized_class(class_name, description, methods)
    
    # 4. Save the Self-Generated Code
    output_file = project_root / "AUTONOMOUS_SYSTEM" / "ai_agent.py"
    output_file.write_text(source_code)
    
    print(f"✅ SUCCESS: System has autonomously written '{output_file.name}'")
    print("   ...File saved with Optimized Async Patterns and Type Hinting.")

if __name__ == "__main__":
    asyncio.run(run_self_construction())
