requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.8
bcrypt==4.0.1
blinker==1.9.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.1.8
ConfigArgParse==1.7.1
coverage==7.10.7
cryptography==46.0.5
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
exceptiongroup==1.3.1
fastapi==0.109.0
filelock==3.19.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.2.5
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.34.0
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
mypy==1.8.0
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
passlib==1.7.4
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
ruff==0.1.14
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.35.1
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.0
virtualenv==20.38.0
Werkzeug==3.1.6
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
